requirements.txt
biopython==1.80